requirements.txt
colorthief==0.2.1
MinePI==0.4.2
Pillow==9.2.0
bokeh
streamlit-image-crop
click==8.1.3